chatrelated/requirements.txt
flask
flask_sqlalchemy
python-dotenv
flask-socketio
jwt
hashlib